requirements.txt
google-generativeai
pytest
python-dotenv
requests
deepdiff
tiktoken